+++ CHANGELOG.md
@@ -21,7 +21,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 --- -->
 
-## [0.1.1] - 2026-02-29
+## [0.1.3] - 2026-03-01
+
+### Added
+
+- Add unit tests for QuickAddForm and EditInventoryItemForm components
+- Add aria-label to form for accessibility improvements
+
+### Changed
+
+- Update quantity labels in AddInventoryItemForm and MainContainer tests for consistency
+- Update quantity labels in AddInventoryItemForm and EditInventoryItemForm for clarity
+- Refactor MainContainer tests to improve form handling and add conditional logic for form submission
+
+---
+
+## [0.1.2] - 2026-03-01
+
+### Fixed
+
+- Add required attribute to Location input in EditInventoryItemForm
+
+---
+
+## [0.1.1] - 2026-03-01
 
 ### Added